Fixed buffers all getting garbage values

+++ SokolFountain/ParticleSystem.cpp
@@ -14,7 +14,7 @@ void ParticleSystem::Tick(float deltaTime, hmm_mat4 params) const
     }
 }
 
-void ParticleSystem::AddEmitter(std::vector<vertex_t>* vertices, std::vector<uint32_t>* indices)
+void ParticleSystem::AddEmitter(std::vector<vertex_t>* vertices, std::vector<uint16_t>* indices)
 {
     emitters.emplace_back(vertices, indices, 2.0f, 4.0f, 512);
 }
